[FIX] #0000 - Alteração para permitir gerar grid do relatório manualmente, configurada via paramentro no optins
